+++ .gitignore
@@ -9,6 +9,10 @@ __pycache__/
 /_skbuild
 build
 
+src/sora_sdk/*.so
+src/sora_sdk/*.dll
+src/sora_sdk/model_coeffs/*
+!src/sora_sdk/model_coeffs/.gitkeep
+
 # パッケージ
 /dist
-MANIFEST.in
