+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include run.py
+include pypath.py
