+++ requirements-dev.lock
@@ -9,6 +9,7 @@
 -e file:.
 absl-py==1.4.0
 attrs==23.1.0
+build==0.10.0
 cffi==1.15.1
 contourpy==1.0.7
 cycler==0.11.0
@@ -18,6 +19,7 @@ importlib-resources==5.12.0
 kiwisolver==1.4.4
 matplotlib==3.7.1
 mediapipe==0.10.0
+nanobind==1.4.0
 numpy==1.24.3
 opencv-contrib-python==4.7.0.72
 opencv-python==4.7.0.72
@@ -27,7 +29,11 @@ pillow==9.5.0
 protobuf==3.20.3
 pycparser==2.21
 pyparsing==3.0.9
+pyproject-hooks==1.0.0
 python-dateutil==2.8.2
 six==1.16.0
 sounddevice==0.4.6
+tomli==2.0.1
 zipp==3.15.0
+# The following packages are considered to be unsafe in a requirements file:
+setuptools==67.8.0
